Move files in the appointment book web original project to the "apptbookweb" package so that we can have unique package names in the original projects that will not confuse the aggregate javadoc build.
