+++ projects-parent/originals-parent/apptbook-web/pom.xml
@@ -158,7 +158,7 @@
         <configuration>
           <archive>
             <manifest>
-              <mainClass>edu.pdx.cs410J.original.Project4</mainClass>
+              <mainClass>edu.pdx.cs410J.apptbookweb.Project4</mainClass>
             </manifest>
           </archive>
         </configuration>
