+++ projects-parent/originals-parent/apptbook-web/src/it/java/edu/pdx/cs410J/apptbookweb/IndexDotHtmlIT.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.apptbookweb;
 
 import edu.pdx.cs410J.web.HttpRequestHelper;
 import edu.pdx.cs410J.web.HttpRequestHelper.Response;
