+++ projects-parent/originals-parent/apptbook-web/src/it/java/edu/pdx/cs410J/apptbookweb/Project4IT.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.apptbookweb;
 
 import edu.pdx.cs410J.InvokeMainTestCase;
 import edu.pdx.cs410J.web.HttpRequestHelper;
