+++ projects-parent/originals-parent/apptbook-web/src/main/java/edu/pdx/cs410J/apptbookweb/AppointmentBookServlet.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.apptbookweb;
 
 import com.google.common.annotations.VisibleForTesting;
