+++ projects-parent/originals-parent/apptbook-web/src/main/java/edu/pdx/cs410J/apptbookweb/Messages.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.apptbookweb;
 
 /**
  * Class for formatting messages on the server side.  This is mainly to enable
