+++ projects-parent/originals-parent/apptbook-web/src/test/java/edu/pdx/cs410J/apptbookweb/AppointmentBookServletTest.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.original;
+package edu.pdx.cs410J.apptbookweb;
 
 import org.junit.Test;
